get rid od utils_precindep
